- Changed Return period to norm - Set actual values for AssessmentSectionComposition Related to Issue [WTI-1019]
git-svn-id: https://repos.deltares.nl/repos/WettelijkToetsInstrumentarium/trunk@6940 5a3db67b-de53-47b5-99c8-a1c30a6650e2

Former-commit-id: db33e630712b9aae8047588a8279b595f0b171be
Former-commit-id: 2446d2799f8d3c5cdba4e78742e4b036feadcff0

+++ Application/Ringtoets/src/Application.Ringtoets.Storage/DbContext/AssessmentSectionEntity.cs
@@ -47,7 +47,7 @@ namespace Application.Ringtoets.Storage.DbContext
         public string Id { get; set; }
         public string Name { get; set; }
         public string Comments { get; set; }
-        public int Norm { get; set; }
+        public double Norm { get; set; }
         public string HydraulicDatabaseVersion { get; set; }
         public string HydraulicDatabaseLocation { get; set; }
         public short Composition { get; set; }
